+++ js-miniapp-sample/src/pages/event-listener.js
@@ -47,6 +47,7 @@ const useStyles = makeStyles((theme) => ({
     width: '90%',
     marginTop: 10,
     marginBottom: 10,
+    background: 'white',
   },
 }));