+++ js-miniapp-sample/src/routes.js
@@ -19,6 +19,8 @@ import EventIcon from '@material-ui/icons/CompareArrows';
 import PhotoCamera from '@material-ui/icons/PhotoCamera';
 import SecurityIcon from '@material-ui/icons/Security';
 import NotificationImportantOutlinedIcon from '@material-ui/icons/NotificationImportantOutlined';
+import ShoppingCartIcon from '@material-ui/icons/ShoppingCart';
+import SendIcon from '@material-ui/icons/SendSharp';
 
 import Ads from './pages/ads';
 import AuthToken from './pages/auth-token';
@@ -39,6 +41,8 @@ import WebLocation from './pages/web-location';
 import WindowActions from './pages/window-actions';
 import SecureStorageComponent from './pages/secure-storage';
 import { CloseConfirmAlert } from './pages/app-close-alert';
+import Purchases from './pages/inapp-purchases';
+import UniversalBridge from './pages/universal-bridge';
 
 const homeItem = [
   {
@@ -61,6 +65,12 @@ const appItems = [
     label: 'Fetch Unique ID from SDK',
     navLink: '/fetch_id',
     element: <UuidFetcher />,
+  },
+  {
+    icon: <ShoppingCartIcon />,
+    label: 'InApp Purchases',
+    navLink: '/inapp_purchases',
+    element: <Purchases />,
   },
   {
     icon: <LocationOnIcon />,
@@ -157,6 +167,12 @@ const appItems = [
     label: 'Close Alert',
     navLink: '/close-confirm-alert',
     element: <CloseConfirmAlert />,
+  },
+  {
+    icon: <SendIcon />,
+    label: 'Universal Bridge',
+    navLink: '/universal-bridge',
+    element: <UniversalBridge />,
   },
 ];